backend/requirements.txt
Django==4.2.16
django-admin-autocomplete-filter==0.7.1
django-filter==23.1
djangorestframework==3.15.2
djoser==2.1.0
webcolors==1.11.1
isort==5.13.2
Pillow==9.3.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
flake8==6.0
flake8-isort==6.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
